Refactored the code. Moved the beans to a separate file.
